Bare minimum of test separation.
Files checked in: DataRepo/templatetags/customtags.py DataRepo/tests/test_compositeviews.py DataRepo/tests/test_views.py DataRepo/tests/test_formats.py
